tests + refactor
